+++ backend/package.json
@@ -10,7 +10,7 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.18.1",
-    "mongoose": "^6.6.4",
+    "mongoose": "^6.6.5",
     "nodemon": "^2.0.20"
   },
   "description": ""
