+++ backend/server.js
@@ -12,7 +12,7 @@ const normalizePort = val => {
   }
   return false;
 };
-const port = normalizePort(process.env.PORT || '3001');
+const port = normalizePort(process.env.PORT || '3010');
 app.set('port', port);
 
 const errorHandler = error => {
