+++ frontend/public/logoNewMe2.ico
