+++ frontend/src/pages/default.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+
+
+
+const Default = () => {
+    return (
+        <div>
+
+        </div>
+    );
+};
+
+export default Default;
