+++ frontend/src/styles/navbar.css
