--- src/components/Navigation.js
@@ -1,28 +1,0 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-
-const Navigation = () => {
-    return (
-        <div className="navigation">
-            <ul>
-                <NavLink to="/" className={(nav) => (nav.isActive ? "nav-active" : "")}>
-                    <li>
-                        Accueil
-                    </li>
-                </NavLink>
-                <NavLink to="/about" className={(nav) => (nav.isActive ? "nav-active" : "")}>
-                    <li>
-                        à propos
-                    </li>
-                </NavLink>
-                <NavLink to='/api' className={(nav) => (nav.isActive ? "nav-active" : "")}>
-                    <li>
-                        API
-                    </li>
-                </NavLink>
-            </ul>
-        </div>
-    );
-};
-
-export default Navigation;
