--- src/components/NewObjectif.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-const NewObjectif = () => {
-    return (
-        <div>
-            <div>
-                <button className='button_new_obj'>Se créer un nouvel objectif?</button>
-            </div>
-        </div>
-    );
-};
-
-export default NewObjectif;
