--- src/styles/index.css
@@ -1,23 +1,0 @@
-.bouton_envie {
-    background-color: red;
-}
-
-.form_connection {
-    background-color: red;
-}
-
-.index_accueil {
-    text-align: center;
-}
-
-.navigation {
-    background-color: green;
-}
-
-.page_index {
-    background-color: blue;
-}
-
-.button_new_obj {
-    background-color: yellow;
-}
